pyproject.toml
[project]
name = "dh-ai"
version = "0.1.0"
description = "AI tools for the https://www.dhcode.cn/"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.4",
    "dotenv>=0.9.9",
    "emoji>=2.14.1",
    "gensim>=4.3.3",
    "ipykernel>=6.29.5",
    "jieba>=0.42.1",
    "matplotlib>=3.10.0",
    "memory-profiler>=0.61.0",
    "nltk>=3.9.1",
    "numpy<2",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "playwright>=1.50.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "spacy>=3.8.4",
    "termcolor>=3.0.0",
    "timm>=1.0.14",
    "torchinfo>=1.8.0",
    "torchvision>=0.15.2",
    "transformers>=4.49.0",
    "uv>=0.6.2",
    "wandb>=0.19.7",
]
[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.uv.workspace]
members = ["第二期/week3-gpt2/gpt-2-Pytorch", "第二期/week3-gpt2/hugging-face/performance_test", "ai_interview/projects/text-similarity", "ai_interview/projects/text-similarity/crawler_news"]
